+++ Plugin/src/MapStuff/Flora.cs
@@ -3,10 +3,7 @@ using UnityEngine;
 
 namespace CodeRebirth.MapStuff;
 public class Flora : MonoBehaviour { // Added for potential future implementations
-    public GameObject[] SteelSprigs = null!;
-    public GameObject[] BluntspearA = null!;
-    public GameObject[] BluntspearB = null!;
-    public GameObject[] PeacockPlant = null!;
-    public GameObject[] Staright = null!;
-    public GameObject[] Misc = null!;
+    public GameObject[] Grass = null!;
+    public GameObject[] Desert = null!;
+    public GameObject[] Snowy = null!;
 }