+++ Plugin/src/Util/ContentHandler.cs
@@ -31,44 +31,25 @@ public class ContentHandler<T> where T: ContentHandler<T> {
         }
     }
 
-    
-	protected void RegisterOutsideObjectWithConfig(string[] tags, int width, GameObject prefab, AnimationCurve curve, string configString) {
-		SpawnableOutsideObjectDef outsideMapObjectDef = ScriptableObject.CreateInstance<SpawnableOutsideObjectDef>();
-		outsideMapObjectDef.spawnableMapObject = new SpawnableOutsideObjectWithRarity
-		{
-			spawnableObject = ScriptableObject.CreateInstance<SpawnableOutsideObject>(),
-		};
-		outsideMapObjectDef.spawnableMapObject.spawnableObject.spawnableFloorTags = tags;
-		outsideMapObjectDef.spawnableMapObject.spawnableObject.objectWidth = width;
-		outsideMapObjectDef.spawnableMapObject.spawnableObject.prefabToSpawn = prefab;
-		(Levels.LevelTypes[] vanillaLevelSpawnType, string[] CustomLevelType) = MapObjectConfigParsing(configString);
-
-        Levels.LevelTypes levelTypes = 0;
-        foreach (Levels.LevelTypes levelType in vanillaLevelSpawnType) {
-            levelTypes |= levelType;
-        }
-        MapObjects.RegisterOutsideObject(outsideMapObjectDef, levelTypes, CustomLevelType, (level) => curve);
-	}
-
-    protected (Levels.LevelTypes[] vanillaLevelSpawnType, string[] customLevelType) MapObjectConfigParsing(string configString) {
-        var spawnRateByLevelType = new List<Levels.LevelTypes>();
-        var spawnRateByCustomLevelType = new List<string>();
+    protected string[] MapObjectConfigParsing(string configString) {
+        var levelTypesList = new List<string>();
 
         foreach (string entry in configString.Split(',').Select(s => s.Trim())) {
             string name = entry;
             if (System.Enum.TryParse(name, true, out Levels.LevelTypes levelType)) {
-                spawnRateByLevelType.Add(levelType);
+                levelTypesList.Add(name);
             } else {
                 // Try appending "Level" to the name and re-attempt parsing
                 string modifiedName = name + "Level";
                 if (System.Enum.TryParse(modifiedName, true, out levelType)) {
-                    spawnRateByLevelType.Add(levelType);
+                    levelTypesList.Add(modifiedName);
                 } else {
-                    spawnRateByCustomLevelType.Add(name);
+                    levelTypesList.Add(name);
                 }
             }
         }
-        return (spawnRateByLevelType.ToArray(), spawnRateByCustomLevelType.ToArray());
+
+        return levelTypesList.ToArray();
     }
 
     protected (Dictionary<Levels.LevelTypes, int> spawnRateByLevelType, Dictionary<string, int> spawnRateByCustomLevelType) ConfigParsing(string configMoonRarity) {